fix: address review feedback — move tests, fix flakiness, return drain error
- Move shutdown race tests from shutdown_race_test.go into grain_test.go
  per maintainer request
- Fix empty-block lint error (for range ch {})
- Make dispatcher.drain() return error so callers can act on timeout
- Short-circuit grain deactivation when drain fails (context expired)
- Replace flaky time.Sleep synchronization with signal channels and
  require.Eventually in tests
- Add runtime.Gosched() yield in mailbox CrossWorkerHandoff test to
  reduce CPU burn under -race
